app/requirements.txt
coreapi==2.3.3
Django==4.0
dj-database-url==0.5.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
psycopg2-binary==2.9.2
whitenoise==5.3.0
